Add test script to package.json
There doesn't appear to be a makefile associated with the project, so
I've assumed the 'make test' command previously listed in scripts was
boilerplate and just blown it away.

Using ./node_modules/.bin so that a developer doesn't (necessarily)
need mocha installed globally on their machine before getting started.

This commit makes 'npm test' run in the project root work as expected.