Fix markdown fields losing controls when hidden
Because of how controls are rendered when hidden (shortcutted by
returning null) if you hide a markdown input & reshow it doesn't get the
markdown editor applied to it again. Prior to this commit that was only
happening in `componentDidMount`, which happens the first time a
component is rendered. This change abstracts out the markdown editor
instantiation a bit so that it's easy to add it to the
`componentDidUpdate` lifecycle method as well. Now the component will
re-attach the textarea when a markdown field is reshown after being
hidden.

This also breaks out the `headingCallback()` as it wasn't necessary to
be defined every time `buttonsToAdd()` executed. It also moves the
functionality of `buttonsToAdd()` inline to markdown editor
instantiation as it's no longer complicated enough to justify its own
function.