Markdown respects height config
Told markdown it should respect the `height` option that was already
being managed on the server but not sent to the client. Then passed it
through to the markdown editor so it'd respect it.

Also taught markdown fields how not to be wysiwyg (though they default
to wysiwyg, is that the right behavior? It's the backwards-compatible
behavior at least). Doing this required slightly more work than a usual
option to maintain backcompat but isn't super-scary or anything. Height
is also maintained when wysiwyg is disabled.

Also sending `toolbarOptions` down to the client. It was in the docs but as far
as I can tell it wasn't plumbed through in the React component for
markdown fields. Only supporting `toolbarOptions.hiddenButtons` for now since
that is all that is in the docs.