Allow wysiwyg option to be an object of options
By making this change it allows you to specify editor config options on
a per-field basis. Since even an empty object is truthy none of the code
around checking for field validity needs to change and we can simply
pass the value through. The React component does have to do an
`_isObject` check now but it's a small change.

Had to rework the `getOptions` function to take the `wysiwyg` object
option into account (via `_.defaults`) and then changed references to
the merged result throughout the function. Very little else actually
needed to change though.

I've tested against global options, fields wit no options, and fields
with customized options all in the same model to try and rule out any
pass-by-reference issues and don't see anything. If someone can think of
a sufficiently-complicated test-case I could use that this breaks on I'd
love to see it. This change has me a bit paranoid.

Fixes #525