Refactor Pagination to allow optional expressions
Added options.optionalExpression value to model.find. This allows an additional object to be passed to the find method. This is useful for example, when performing a $text search, and you want to sort by score. In this case, your filter would be: { $text: { $search: term } } and optionalExpression would be: { score: { $meta: 'textScore' } }.
Added an additional query that does not include the optionalExpression object to perform the count of objects. This allows the count to be performed without additional meta data from the optionalExpression interfering.