Pagination Functional Tests
Added simple "Post" model with a multi-key, weighted index.
Added test and assertions for list.paginate to ensure no existing functionality has been affected by the addition of the `optionalExpression` and `countQuery` modifications.
Added test and assertions for a paginated result set that includes an `optionalExpression` that adds metadata to the result.