Ensuring File upload keys contain the field type
This guarantees uniqueness when different field types share the same
path (i.e. with nested Lists) cc/ @wmertens

The previous fix could cause problems because the fully qualified
`inputName` can contain characters that bodyParser extrapolates into
nested data, or an array, and we need these fields to be present in the
top level of the `req.files` object regardless of where they are
referenced in a nested / repeating data structure.