update handler handles having no req
Currently if `getUpdateHandler` is passed no arguments, it will cause
the process to fail in an error, as we later check req.user and
req.files.

This change means we will continue to return undefined unless there is
a req.user ora req.files, which was the previous behaviour.

Additionally, req is set as a variable earlier, so am checking the
function-local variable.