write keystone bundles to disk
this will expose a keystone config option that will allow a developer to
tell keystone to write (admin.js, fields.js, signin.js) to
a file in the static directory

this means that keystone won't need to recompile these files between
server boots. this brings the boot time down from 20s on every launch to
around 3-5s if the files are already compiled and available. this
additional mode is compatible with the previous strategy of storing each
bundle in a variable in the closure which it will still do in order to
enable the watchify workflow.

this mode is enabled by default, and can be disabled by explicitly
setting 'cache admin bundles' to false