Exception with the JSON download & Fix
Hi!

Firstly can I say going through this codebase is a joy, I've known of keystone for awhile and tinkered with it in my spare time - its great to see how its evolved. 

I found that this was throwing an exception when you click to download data as json.

you can reproduce the bug at http://demo.keystonejs.com/keystone/posts download > json > download

the server hangs with no response... 

this fixes the issue, the flow is now similar with the CSV download. Which is where I pinched the 'list-to-array' require from. Its now fixed for this scenario.

A few other points I have noted for discussion:
- Do we also want it to be a json file download or json response?
- we should probably also have an exception handler to return 500 from express for anything else?
- I notice the user can enter "password" into the query string and reveal the hash, is this expected behaviour? I assume its safe under a users logged in session
- I've noticed the require for 'baby-parse' is within the handler? do we want to hang the server on request? https://github.com/keystonejs/keystone/blob/master/admin/server/api/list/download.js#L9

If you're after any more help on issues or more info on this bug, I'd be more than happy to help! 

Thanks 

Mike

## Description of changes

I've added the missing check for "is a string" if so then convert to an array which done by the third-party dependency 'list-to-array'.

## Testing

a note: I got this selenium error before I made changes on the branch, this has actually prevented me from getting the selenium tests to pass. 

```Timed out while waiting for element <.primary-navbar [data-section-label="octicon-sign-out"]> to be present for 60000 milliseconds.  - expected "visible" but got: "not found"
    at Object.after (/Users/MikeJames/projects/keystone-play/keystone/test/e2e/adminUI/tests/group006Fields/commonFieldTestUtils.js:13:22)
```