Added test to cover validation of a custom display format.
Altered test for validation of the default format when a parseFormat
is specified.

Previously, if a parseFormat is specified, this became the only
acceptable format, even if a different display format is defined.
If the display format differs from the parse format, the field will
not validate as populated from the database.

Now, if a parseFormat (or formats, in an array) is specified it will
be added to the list of acceptable formats (the default array plus
the display format).